digisim

+++ os/main.c
@@ -42,7 +42,7 @@ int solve_grid(grid_t *grid, int level) {
       // iterate over all columns
       for (j = 0; j < 9; j++) {
         int box, sol, solcount;
-        int v[10] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+        int8_t v[10] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
 
         // skip slots that already have an assigned value
         if (g.grid[9 * i + j] != 0) {
